Add Prescout drawer screen and reorganize settings
