+++ app/(drawer)/prescout/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from 'expo-router';
+
+export default function PrescoutLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name="index" options={{ title: 'Prescout' }} />
+    </Stack>
+  );
+}
