+++ app/(drawer)/prescout/index.tsx
@@ -1,0 +1,5 @@
+import { PrescoutScreen } from '@/app/screens';
+
+export default function PrescoutRoute() {
+  return <PrescoutScreen />;
+}
