+++ app/screens/Prescout/PrescoutScreen.tsx
@@ -1,0 +1,5 @@
+import { TeamListScreen } from '@/app/screens/Shared/TeamListScreen';
+
+export function PrescoutScreen() {
+  return <TeamListScreen title="Prescout" />;
+}
