+++ app/screens/index.ts
@@ -1,5 +1,6 @@
 export { LoginScreen } from './Auth/LoginScreen';
 export { PitScoutScreen } from './PitScout/PitScoutScreen';
+export { PrescoutScreen } from './Prescout/PrescoutScreen';
 export { MatchScoutScreen } from './MatchScout/MatchScoutScreen';
 export { MatchTeamSelectScreen } from './MatchScout/MatchTeamSelectScreen';
 export { AppSettingsScreen } from './Settings/AppSettingsScreen';
