+++ constants/routes.ts
@@ -2,6 +2,7 @@ export const ROUTES = {
   login: '/auth/login' as const,
   pitScout: '/(drawer)/pit-scout' as const,
   matchScout: '/(drawer)/match-scout' as const,
+  prescout: '/(drawer)/prescout' as const,
   appSettings: '/(drawer)/settings' as const,
   eventsBrowser: '/(drawer)/settings/events' as const,
   userSettings: '/(drawer)/user-settings' as const,
